Rename descriptor for skipped specs
